+++ src/controllers/user.js
@@ -1,0 +1,5 @@
+let user
+
+// TODO create user
+
+module.export = user