+++ src/index.js
@@ -1,4 +1,5 @@
 const express = require('express')
+const dbClient = require('./dbClient')
 
 const app = express()
 const port = 3000
@@ -6,8 +7,14 @@ const port = 3000
 app.get('/', (req, res) => {
   res.send('Hello World!')
 })
- 
-module.exports = app.listen(port, (err) => {
+
+dbClient.on("error", function(error) {
+  console.error(error);
+});
+
+const server = app.listen(port, (err) => {
   if (err) throw err
   console.log("Server listening the port " + port)
 })
+
+module.exports = server